requirements.txt
streamlit
tensorflow
numpy
Pillows